anno_env/requirements.txt
opencv-python==4.8.0.74
matplotlib
numpy
pillow
pygame
pykalman
pyyaml
pandas
open3d==0.18.0
scipy
statsmodels==0.13.2
tqdm
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
anaconda-client==1.2.2
anyio==3.5.0
appdirs==1.4.4
argh==0.26.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==0.13.1
asn1crypto==1.4.0
astroid==2.6.6
astropy==5.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.1
